- Merged changes from 1.20.1

+++ src/main/java/com/ilmusu/musuen/callbacks/PlayerPhantomSpawnCallback.java
@@ -14,5 +14,5 @@ public interface PlayerPhantomSpawnCallback
                 return insomniaAmount;
             });
 
-    int handler(PlayerEntity player, int insomniaAmount);
+    float handler(PlayerEntity player, float insomniaAmount);
 }
